Anchor the decision grid on the clock, and tell the spec builder which label it is building
Three RoboRev HIGH findings on the previous push, jobs #17876 and #17877. All three were
right and the first changes the design.

**The grid anchors on the clock, not on the panel's first row.** Anchoring on the first
available timestamp put this panel's decisions at 10:31, 10:46, 11:01, because the feature
matrix starts after the warmup hour the features pay. `03_financial_features.py` already
states the decision grid as `minute % DECISION_MINUTES == 0` and `17_costs.py`'s default
arm reads the same one, so the session anchor disagreed with every other statement of the
grid in the case study. On the clock, a fifteen-minute cadence decides on the quarter hour
in every session, whatever time the panel starts and however early the session closes - a
short session simply holds fewer decisions rather than a different set of them.

**`apply_rebalance_step` is gone with it.** It existed to stop a multi-slot step drifting
across the overnight boundary, which the constructed schedule now prevents at the source:
nothing intraday needs a step above 1 once the cadence carries the horizon. Keeping it
would have changed `crypto_perps_funding`, whose `fwd_ret_24h` steps 3 slots of an
8-hourly panel and which is done. The step application is back to the `gather_every` it
always was.

**The four nasdaq callers pass `label=`.** Declaring `decision.cadence_by_label` activates
`build_backtest_spec`'s mandatory-label check, and `14_backtest` (both sites),
`15_portfolio_management` and `17_costs` omitted it, so every one of them would have
raised on its first cell. Each already has `LABEL` in scope and already passes it to
`run_backtest`; only the spec builder was missing it.

Verification: `tests/test_intraday_decision_schedule.py` pins the clock alignment, the
early-close case, the sweep arms separating on their token alone, and - because that case
study is done and may not move - an 8-hourly panel under its own cadence resolving to
itself across a missing funding time. Against this branch's base, the affected suites go
118 passed to 129 passed with the same three pre-existing failures, all of which are
missing local market data rather than behaviour.
