+++ case_studies/utils/backtest_loaders.py
@@ -1356,17 +1356,26 @@ def resolve_rebalance_timestamps(
         # decisive and makes a panel at any resolution produce the same schedule - which is
         # what the Ch18 cadence sweep needs, since its arms differ only in this token.
         #
-        # Anchored on each session's first available timestamp, not on midnight: the panel
-        # starts when the features finish warming up (10:31 for nasdaq100), and an offset
-        # from midnight would put the first decision of every session at a different lag
-        # from the data it can actually see. Sessions are also not all the same length, so a
-        # count that runs across the close walks to a different time of day and stays there.
+        # Anchored on the clock, not on the session's first row. That is the grid
+        # `03_financial_features.py` already calls the decision grid (`minute % 15 == 0`),
+        # and it is the one a reader can name: a fifteen-minute cadence decides on the
+        # quarter hour, every session, whatever time the panel happens to start at and
+        # however early the session closes. Anchoring on the first available row instead
+        # would put the decisions of a panel that starts at 10:31 - which this one does,
+        # after the warmup hour the features pay - at 10:31, 10:46, 11:01, agreeing with no
+        # other statement of the grid in the case study.
         seconds = int(interval.total_seconds())
         return (
             pl.DataFrame({"ts": ts})
-            .with_columns(_d=pl.col("ts").dt.date())
-            .with_columns(_o=(pl.col("ts") - pl.col("ts").min().over("_d")).dt.total_seconds())
-            .filter(pl.col("_o") % seconds == 0)
+            .filter(
+                (
+                    pl.col("ts").dt.hour().cast(pl.Int64) * 3600
+                    + pl.col("ts").dt.minute().cast(pl.Int64) * 60
+                    + pl.col("ts").dt.second().cast(pl.Int64)
+                )
+                % seconds
+                == 0
+            )
             .get_column("ts")
         )
 
@@ -1448,37 +1457,6 @@ def resolved_rebalance_step(rebalance_spec: dict | None, case_study: str, label:
     return get_rebalance_step(case_study, label)
 
 
-def apply_rebalance_step(schedule: pl.Series, step: int) -> pl.Series:
-    """Keep every ``step``-th scheduled slot, anchored per session when the schedule is intraday.
-
-    A plain ``gather_every`` over the whole schedule is right for a daily-or-coarser grid,
-    where one slot is one session and sessions are what the step counts. It is wrong for an
-    intraday grid, because the count then runs across the overnight boundary: a session whose
-    length is not a multiple of the step shifts the phase of every session after it, and the
-    shift accumulates.
-
-    NASDAQ regular sessions hold 390 minutes and early closes 210. Both divide by 5 and by 15,
-    so those steps survive a global gather by coincidence; 60 divides neither, so an hourly
-    cadence would start 30 minutes later after every early close and never come back. Anchoring
-    per session makes the first decision of each session its first slot, which is what a
-    declared intraday cadence means and what a reader checking one day against another expects.
-
-    A schedule with at most one timestamp per calendar date is not intraday, and is gathered
-    globally so daily, weekly and monthly cadences keep the identity they already have.
-    """
-    if step <= 1 or schedule.is_empty():
-        return schedule
-    dates = schedule.dt.date()
-    if dates.n_unique() == schedule.len():
-        return schedule.gather_every(step)
-    return (
-        pl.DataFrame({"ts": schedule})
-        .with_columns(_d=pl.col("ts").dt.date())
-        .filter(pl.int_range(pl.len()).over("_d") % step == 0)
-        .get_column("ts")
-    )
-
-
 def declared_rebalance_step(case_study: str, label: str) -> int | None:
     """The declared step for ``(case_study, label)``, or None when none is declared.
 
@@ -1537,9 +1515,9 @@ def thin_to_rebalance_dates(
     # Step 1: Calendar-aware schedule resolution
     schedule_dates = resolve_rebalance_timestamps(all_dates, cadence)
 
-    # Step 2: Apply design-time non-overlapping step, anchored per session on an
-    # intraday schedule (see apply_rebalance_step).
-    schedule_dates = apply_rebalance_step(schedule_dates, step)
+    # Step 2: Apply design-time non-overlapping step
+    if step > 1:
+        schedule_dates = schedule_dates.gather_every(step)
 
     # Semi-join to filter — avoids Polars is_in precision mismatch
     # (group_by().agg(max) can change Datetime precision)
