+++ case_studies/utils/backtest_runner.py
@@ -1398,7 +1398,6 @@ def _run_engine(
     # ~step× too rarely. The on_data callback already gates on
     # ``timestamp in weight_dict``, so dates without weights are skipped.
     from case_studies.utils.backtest_loaders import (
-        apply_rebalance_step,
         resolve_rebalance_timestamps,
         resolved_rebalance_step,
     )
@@ -1408,7 +1407,8 @@ def _run_engine(
     schedule_dates = resolve_rebalance_timestamps(all_pred_ts, cadence, calendar)
     if case_study and label:
         step = resolved_rebalance_step(rebalance_spec, case_study, label)
-        schedule_dates = apply_rebalance_step(schedule_dates, step)
+        if step > 1:
+            schedule_dates = schedule_dates.gather_every(step)
     rebalance_schedule = {
         _engine_timestamp(
             timestamp,
